ENH: pairPotentialList: make run with collated file format.
pairPotentialList writes some raw tables (writeEnergyAndForceTables) to the
local processor directories (which don't exist in case of collated).
The solution is to create these processor directories. If there
would be some use for these files to be re-read we could
wrap them up in some kind of container so they'd get collated.